feat[api]: add cookie-parser and uuid dependencies and update problem solution
Added the cookie-parser and uuid dependencies to the package.json and the lock file. Created a new dto for the Fingerprint and renamed the addRecord and cacheState dtos. The controller was updated to use the new dependencies and added records and calculated mean now do so by the fingerprint, ruling out the need for a solution in the request body. The service was updated with new methods to match the controller changes. Initialized the app to use the cookieParser.