feat: Improve logging and error handling in API requests, refine cache state management
This commit introduces several major changes to improve the application's functionality. First, it improves logging and error handling for POST and GET requests in the 'records' and 'mean' route files, providing comprehensive logs for the complete lifecycle of a request.

It also introduces a refined cache state management process that ensures the cache state gets persisted to the database correctly, preventing data corruption or loss. This includes converting a deque to an array before saving and converting it back to a deque after retrieval.

Another change involves validating the presence of a fingerprint in request cookies rather than generating a new one if not found. This is done to better identify and resolve issues related to missing fingerprints.

A minor change of multiplying ttl value by 1000 has been done in 'page.tsx' to convert seconds to milliseconds to match the requirements of the initialisation function.

Lastly, there is the removal of unused 'double-ended-queue' and '@types/double-ended-queue' package from 'pnpm-lock.yaml'.