feat: add classnames package and improve UI
Added the 'classnames' package and used it to improve the user interface. The changes include a remodeling of checkboxes and related labels for better user experience. This involves better visual toggling and positioning of labels, improving appearances and dynamics of checkboxes. Also, consequential CSS changes are made to complement this improvement.