feat: add kafka integration and events tracking
The commit includes the creation of KafkaModule and the addition of KafkaJS in the api application. It enhances the MeanCacheProblemService by tracking events and sending messages to a Kafka topic whenever any state is changed. The Kafka producer is configured as a global provider.

These changes also include the addition of Zookeeper, Redpanda, and console services in docker-compose for creating a local Kafka setup.