feat: add kafka topic viewer and update related services
Updated `forward-ports.sh` script to include the port-forwarding command for the Kafka Topic Viewer.

In the `kafka-topic-viewer.yaml` file, image definition is updated to use the vectorized/console image and the configuration is updated to enable schemaRegistry and adminApi for communication with kafka-controller service.

Updated `kafka-consumer.yaml` bootstrap server's value to reflect the Kafka service in the default namespace.